Added more information to GetOne project with ID

+++ src/service/controllers/project_enroll.go
@@ -38,7 +38,7 @@ func (c *UserEnrollOnProjectController) Post() {
 
 	} else {
 		// получить id проекта, на который пользователь хочет записаться
-		project_id, err := c.GetInt64(":id")
+		project_id, err := c.GetInt(":id")
 		if err != nil {
 			beego.Debug(c.Ctx.Input.IP(), "Not an int param. Should be int", err.Error())
 			c.Ctx.Output.SetStatus(HTTP_BAD_REQUEST)
