+++ src/service/controllers/project_master.go
@@ -37,7 +37,7 @@ func (c *ProjectMasterController) Post() {
 		c.Data["json"] = HTTP_FORBIDDEN_STR
 
 		// получить id проекта, на который нужно добавить мастера
-	} else if project_id, err := c.GetInt64("project_id"); err != nil {
+	} else if project_id, err := c.GetInt("project_id"); err != nil {
 		beego.Debug("Not an int param. Should be int", err.Error())
 		c.Ctx.Output.SetStatus(HTTP_BAD_REQUEST)
 		c.Data["json"] = err.Error()
@@ -104,7 +104,7 @@ func (c *ProjectMasterController) GetOne() {
 		c.Ctx.Output.SetStatus(HTTP_BAD_REQUEST)
 		c.Data["json"] = err.Error()
 
-	} else if v, err := models.GetMastersOfTheProject(int64(id)); err != nil {
+	} else if v, err := models.GetMastersOfTheProject(id); err != nil {
 		beego.Debug("GET masters of project error", err.Error())
 		c.Data["json"] = err.Error()
 		c.Ctx.Output.SetStatus(HTTP_BAD_REQUEST)
@@ -186,7 +186,7 @@ func (c *ProjectMasterController) Delete() {
 		c.Ctx.Output.SetStatus(HTTP_BAD_REQUEST)
 		c.Data["json"] = "You can't delete yourself"
 
-	} else if project_id, err := c.GetInt64("project_id"); err != nil {
+	} else if project_id, err := c.GetInt("project_id"); err != nil {
 		beego.Debug("Param `user_id` not int", err.Error())
 		c.Ctx.Output.SetStatus(HTTP_BAD_REQUEST)
 		c.Data["json"] = err.Error()
